COMP: more lex.yy.c warnings

+++ Wrapping/lex.yy.c
@@ -1573,8 +1573,8 @@ static int input (void );
         if ( YY_CURRENT_BUFFER_LVALUE->yy_is_interactive ) \
                 { \
                 int c = '*'; \
-                int n; \
-                for ( n = 0; n < max_size && \
+                size_t n; \
+                for ( n = 0; n < (size_t)max_size && \
                              (c = getc( yyin )) != EOF && c != '\n'; ++n ) \
                         buf[n] = (char) c; \
                 if ( c == '\n' ) \
@@ -1586,7 +1586,7 @@ static int input (void );
         else \
                 { \
                 errno=0; \
-                while ( (result = fread(buf, 1, max_size, yyin))==0 && ferror(yyin)) \
+                while ( (result = fread(buf, 1, (size_t)max_size, yyin))==0 && ferror(yyin)) \
                         { \
                         if( errno != EINTR) \
                                 { \
@@ -2861,7 +2861,7 @@ static int yy_get_next_buffer (void)
 
         else
                 {
-                        size_t num_to_read =
+                        int num_to_read =
                         YY_CURRENT_BUFFER_LVALUE->yy_buf_size - number_to_move - 1;
 
                 while ( num_to_read <= 0 )
